Move q!lcc to /lcc
